realtime-digit-recognition/requirements.txt
Flask==3.1.1
numpy==2.1.3
pillow==11.3.0
tensorflow==2.20.0
